### 2020-11-23 #92 @whedon
[PRE REVIEW]: ParFlow Sand Tank: A tool for groundwater exploration
**Submitting author:** @lecondon (<a href="http://orcid.org/0000-0003-3639-8076">Laura Condon</a>)
**Repository:** <a href="https://github.com/hydroframe/SandTank" target ="_blank">https://github.com/hydroframe/SandTank</a>
**Version:** v1.0.6

@lecondon if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
