### 2020-11-23 #92 @labarba
hi @lecondon — thank you for your submission to JOSE. We've had the journal in a holding pattern since the pandemic hit, mainly because I became overwhelmed with pivoting online teaching, research, and JupyterCon (I was General Chair). Thank you for your patience. I am beginning to dig up from a long backlog, so you'll begin seeing some movement.

Could you help with some leads of people we might reach out to for a review of this work? Many thanks!
