### 2020-11-24 #92 @labarba
@marda-science has agreed to be the first reviewer for this submission—thank you!!

We'll be looking for a second reviewer, and then start the REVIEW issue, where all the action happens. Meanwhile, you could have a look at an [old example review](https://github.com/openjournals/jose-reviews/issues/53): you'll get a checklist as shown there to guide you in the review, and you'll be encouraged to open issues on the submission repository (adding a cross-link with the review). Feel free to start inspecting the submission in the target repository: https://github.com/hydroframe/SandTank
