### 2020-12-04 a13e19d
fix(Indicator): Color mapping
