### 2021-01-04 #92 @lecondon
Hi Lorena,

No worries about the timing. I’m clearly a bit behind too.

Are you still looking for a second reviewer? Here’s one name. I can come up with more too if you need.

- Nick Engdahl nick.engdahl@wsu.edu<mailto:nick.engdahl@wsu.edu>

Thanks,

Laura

—
Laura E. Condon, PhD
Assistant Professor
Hydrology and Atmospheric Sciences
University of Arizona, Tucson, AZ
