### 2021-01-05 #92 @lecondon
Great thanks Bryan! Let me know if you need additional suggestions.

Laura

—
Laura E. Condon, PhD
Assistant Professor
Hydrology and Atmospheric Sciences
University of Arizona, Tucson, AZ
