### 2021-04-22 ac0e562
Allow total pumping to sum until SandTank reset
