### 2021-04-22 9f6ddb5
Publish total storage and total pumping as single values
