### 2021-07-13 #92 @labarba
@DamienIrving – hi! 👋  You have published with JOSE in the past, and I wonder if you would consider now contributing a review. We are in need of some help with this submission:

Title: **ParFlow Sand Tank: A tool for groundwater exploration**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00092/jose.00092/10.21105.jose.00092.pdf)

Thanks for considering it!
