### 2021-07-21 bab4804
Make new features toggleable
