### 2021-07-21 d46bd06
Isolate new features to hillslope_s template and make them toggleable
