+++ deploy/pvw/server/api.py
@@ -111,9 +111,24 @@ class SandTankEngine(pv_protocols.ParaViewWebProtocol):
             f.write('set hleft          %s\n' % self.lastConfig['hLeft'])
             f.write('set hright         %s\n' % self.lastConfig['hRight'])
             f.write('\n')
-            f.write('set recharge       %s\n' % self.lastConfig['recharge'])
-            f.write('set waterUseEfficiency         %s\n' % self.lastConfig['waterUseEfficiency'])
-            f.write('set irrigationEfficiency       %s\n' % self.lastConfig['irrigationEfficiency'])
+
+            # If optional values of 'recharge', 'waterUseEfficiency', or 'irrigationEfficiency' are not
+            # found in previous config file, set default values
+            if 'recharge' in self.lastConfig:
+                f.write('set recharge       %s\n' % self.lastConfig['recharge'])
+            else:
+                f.write('set recharge       0\n')
+
+            if 'waterUseEfficiency' in self.lastConfig:
+                f.write('set waterUseEfficiency         %s\n' % self.lastConfig['waterUseEfficiency'])
+            else:
+                f.write('set waterUseEfficiency         1\n')
+
+            if 'irrigationEfficiency' in self.lastConfig:
+                f.write('set irrigationEfficiency       %s\n' % self.lastConfig['irrigationEfficiency'])
+            else:
+                f.write('set irrigationEfficiency       1\n')
+
             f.write('\n')
             f.write('set lake           %s\n' % self.lastConfig['isLake'])
 
