### 2021-07-26 68f3a1b
fix(ci): Update docker build
