### 2021-07-26 1bfc49e
fix(ci): Correct docker build fix
