### 2021-11-02 #92 @leouieda
Hi @bryanwweber thank you for the invitation! I'm a bit hesitant to accept since my time is very limited at the moment and I'm not familiar with a majority of the software stack used and only tangentially aware of the subject matter. Under ideal conditions I would welcome the chance to learn more about docker and JS but I'm afraid that I would not have the time to invest at the moment. So my contribution may not be that useful and I would fear holding back the review process.
