### 2022-03-31 #92 @labarba
@lecondon  — @Bryan Weber has started a new job in Coiled and has decided to step down from the JOSE board, as he does not think that he will be able to carve out time to volunteer with us. We will be re-assigning this submission to a new editor soon. Thank you for your patience! 🙏
