### 2022-06-22 #92 @djlampert
I can try to review this, along with my student @bjrubinstein. I have a background in hydrologic models.
