### 2022-07-25 hydroframe/SandTank#61 @bjrubinstein
Question about authorship // JOSE REVIEW
@lecondon I'm reviewing your submission for JOSE and I have a question about the author's list. I was looking through the commits and it looks like @callachennault has made significant contributions to the code but is not listed as an author and Lisa Gallagher is listed as an author but doesn't have any public contributions to the repo. Can you please clarify this for me?
