### 2022-07-25 hydroframe/SandTank#61 @reedmaxwell
Hi @bjrubinstein thanks for reviewing this manuscript for JOSE!  I'm answering for @lecondon who is out on maternity leave.  I'm not sure that GitHub reflects her email OoO.  The paper was written and submitted more than two years ago and this was before @callachennault started on the project.  My guess is that it also has my old affiliation, I changed institutions a bit more than two years ago as well.  I'm sorry about the confusion, I had completely forgotten that we submitted this manuscript and had more or less given up on the journal.  I'm delighted that review might actually occur and we are happy to add @callachennault as an author.  (Dr.) Lisa Gallagher is an education researcher and specialist and an important contributor to this work.  She has contributed substantially to the design of the software package and I suspect that most productive educational software packages have non-research software engineers who contribute to the project.  Hope this helps clarify, thanks much.
