### 2022-07-25 hydroframe/SandTank#61 @djlampert
So, it seems like the affiliations for the article need to be updated, then we can move forward with the review for JOSE?
