### 2022-07-25 hydroframe/SandTank#62 @djlampert
Question about release version // JOSE REVIEW
Hello, I am also reviewing this article/repo for the Journal of Open Source Education (JOSE) along with @bjrubinstein. The release version listed in the manuscript is 1.0.6, but the release for the GitHub repository is 1.1.1. Please update. Thanks.
