### 2022-07-26 #179 @bjrubinstein
@JasonJWilliamsNY The software version and author list (additional contributor and some changed affiliations) was accurate at the time it was submitted but have changed since the paper was in the review queue for so long. Should these be updated to reflect the current state of the project?
