### 2022-07-26 hydroframe/SandTank#61 @bjrubinstein
@reedmaxwell Thanks for the clarifications. I'm not sure if they expect things to be updated or if it's acceptable that they were accurate at the time the paper was submitted.

I left a question for the editor in the review issue: https://github.com/openjournals/jose-reviews/issues/179#issuecomment-1196105835 I'll update when we hear back.
