### 2022-07-26 hydroframe/SandTank#63 @bjrubinstein
Fix DOI for OseiK:2014 in BibTeX file // JOSE REVIEW
@reedmaxwell The DOI field for Improved numerical solvers for implicit coupling of subsurface and overland flow by Osei-Kuffuor, D. and Maxwell, R. M. and Woodward, C. S. includes a "https://doi.org/" prefix.

Current:
`DOI = {https://doi.org/10.1016/j.advwatres.2014.09.006},`

Should be:
`DOI = {10.1016/j.advwatres.2014.09.006},`
