### 2022-07-27 hydroframe/SandTank#64 @bjrubinstein
Node.js version dependency documentation // JOSE REVIEW
@reedmaxwell The instructions for building the web client in the readme or running the development web server in the Web Application page of the documentation have the following commands respectively:

```
cd ./client
npm install
npm run build
```

```
cd client
npm run serve
```

These fail with the a ERR_OSSL_EVP_UNSUPPORTED on Node version 18 or higher. Last time the documentation was updated version 18 hadn't been released but now that the dependency exists could you please update the documentation to reflected it?
