### 2022-08-04 #179 @bjrubinstein
@lecondon @reedmaxwell @JasonJWilliamsNY @djlampert

I've completed my review, and have only a few minor recommendations. The linked issues have more details.

Add @callachennault to the authors list and update affiliations for the current authors. https://github.com/hydroframe/SandTank/issues/61

Update the manuscript version to match the current Github release version (1.1.1) https://github.com/hydroframe/SandTank/issues/62

Correct a DOI formatting issue in the BibTeX file. https://github.com/hydroframe/SandTank/issues/63

Document the Node.js version dependency. https://github.com/hydroframe/SandTank/issues/64

Other than those few issues everything worked well. I hope this is helpful.
