### 2022-08-08 hydroframe/SandTank#64 @jourdain
A quick google search lead me to [`export NODE_OPTIONS=--openssl-legacy-provider`](https://stackoverflow.com/questions/69394632/webpack-build-failing-with-err-ossl-evp-unsupported). Not sure if that fix it or not.
