+++ paper/paper.bib
@@ -56,7 +56,7 @@
    volume = {74},
    pages = {185-195},
    ISSN = {0309-1708},
-   DOI = {https://doi.org/10.1016/j.advwatres.2014.09.006},
+   DOI = {10.1016/j.advwatres.2014.09.006},
    url = {http://www.sciencedirect.com/science/article/pii/S0309170814001894},
    year = {2014},
    type = {Journal Article}
@@ -81,7 +81,7 @@
    volume = {6},
    number = {3},
    pages = {233-247},
-   DOI = {doi:10.1175/JHM422.1},
+   DOI = {10.1175/JHM422.1},
    year = {2005},
    type = {Journal Article}
 }
