### 2022-08-16 #179 @lecondon
@bjrubinstein @djlampert thank you for your reviews!

I have completed two comments and have questions on two:
Completed comments:
1. We have added Calla to the author list and closed that issue
2. I fixed the DOI and closed that issue as well

Questions:
1. We are fine with updating the version number but I don't see where to go to do this. I left a response noting this in the issue on the repo. If you can clarify what we should do happy to address this one.
2. I see there was some back and forth on the NodeJS issue and it looks like you got it working by updating. Is this resolved or are there changes you would like to see?
