### 2022-08-17 #179 @bjrubinstein
@lecondon Thanks for addressing those issues. Regarding the questions:

1. @JasonJWilliamsNY can you please help us with this? It seems like updating the version in the submission is something that needs to be done from your side. The issue is that the version listed here in the first post is 1.0.6 but the current version in their repo is 1.1.1. Everything in their repo is correct, it is only the version listed here that needs to be updated.
2. I would recommend adding a small note somewhere that says that building the client or running the development server requires Node.js version 17 or lower. The two places this comes up are the readme and the Web Application documentation so maybe mention it in one or both of those places (see the issue for the exact commands that cause the problem).
