### 2022-08-24 #179 @JasonJWilliamsNY
@lecondon any outstanding changes? @djlampert I know you and @bjrubinstein co-reviewed and so probably went over details together but can you confirm things on your checklist in this issue so we can go to the next stages? Thanks
