### 2022-08-25 2c69954
Adding a nodejs version note to the readme
