### 2022-08-25 #179 @lecondon
@djlampert I just added a note to the readme for the node js version. I think this was the only thing left outstanding unless @bjrubinstein  has anything to add.  Thanks
