### 2022-09-14 #179 @djlampert
Ok, it looks like I was able to update the article pdf, and I can see the Statement of Need. I think we have completed reviewing the checklist. I am not sure what else we need to do here. Thanks.
