### 2022-10-03 #179 @lecondon
Okay it looks like this was actually already enabled through zenodo and the doi is: 10.5281/zenodo.5138519. Let me know if you need anything else.

![image](https://user-images.githubusercontent.com/12488567/193493448-8601ecff-e057-4874-a0cb-9f7cfba9e504.png)
