### 2022-10-06 #179 @JasonJWilliamsNY
@lecondon - Could you edit the metadata on the Zenodo deposit so the title and author list matches the JOSE paper?
On the record page and click the orange Edit button in the top-right corner. You will be taken to a form similar to the upload form, where you will be able to edit almost all of the record's metadata. Once you are done modifying it, click "Save" and then "Publish". (See https://help.zenodo.org/)
