### 2022-10-11 #92 @lecondon
HI Jason,

Thank you for your email. Yes to be honest we had pretty much given up on this. If it can proceed that’s great though. I will be on maternity leave after June 30th but I think my co-authors could probably handle the revisions.

Thanks,

Luara

From: Jason Williams ***@***.***>
Date: Thursday, June 2, 2022 at 8:43 AM
To: openjournals/jose-reviews ***@***.***>
Cc: Condon, Laura - (lecondon) ***@***.***>, Mention ***@***.***>
Subject: [EXT]Re: [openjournals/jose-reviews] [PRE REVIEW]: ParFlow Sand Tank: A tool for groundwater exploration (#92)

External Email

Sent via email as well:

Hello Dr. Condon,

I am a new(-ish) editor at the Journal for Open Source Education. I am on assignment to clear out submissions that had been unfortunately backlogged due to COVID. I wanted to reach out by email (can be hard to keep track of GitHub notifications) to confirm you’d like to continue publication of this with JOSE. If so, I wanted to see that the timing is right on your end as I recruit reviewers so that the process can be as smooth as possible once review begins.

Please let me know what you think and I can follow up.

Thanks,

  *   Jason
