### 2022-12-01 6e7b47e
chore(release): 1.1.4
## [1.1.4](https://github.com/hydroframe/SandTank/compare/v1.1.3...v1.1.4) (2022-12-01)

### Bug Fixes

* **zenodo:** token should work now ([b8d23a8](https://github.com/hydroframe/SandTank/commit/b8d23a8526f550da139146e997a7efbc59bb3aee))
