### 2022-12-01 043f005
chore(release): 1.1.6
## [1.1.6](https://github.com/hydroframe/SandTank/compare/v1.1.5...v1.1.6) (2022-12-01)

### Bug Fixes

* **zenodo:** change license to "other-open" ([26bbc32](https://github.com/hydroframe/SandTank/commit/26bbc32f92b7b0b4e704103a4d1c59dd88fc8801))
