### 2022-12-01 #179 @lecondon
It turns out that those instructions didn't work because the CI was pushing to zenodo. I think we have it all updated correctly now though so we should be good to go? Let us know if you need anything else.
