### 2023-03-23 #179 @labarba
Congratulations, @lecondon, your JOSE paper is published! 👏

Huge thanks to the Editor: @JasonJWilliamsNY, and Reviewers: @djlampert, @bjrubinstein — we are grateful for your contributions 🙏
